api/requirements.txt
django==3.2.6
PyJWT==2.1.0
django-utils-six==2.0.0
graphene-django==3.0.0
requests==2.26.0
bcrypt==3.2.0
django-cors-headers==2.0.0